+++ app/lib/service/csp/default_csp.dart
@@ -64,8 +64,10 @@ final _defaultContentSecurityPolicyMap = <String, List<String>>{
   ],
   'connect-src': <String>[
     "'self'",
-    'https://www.google-analytics.com/',
-    'https://stats.g.doubleclick.net/',
+    'https://*.google-analytics.com',
+    'https://*.analytics.google.com',
+    'https://*.googletagmanager.com',
+    'https://*.g.doubleclick.net',
   ],
   'frame-ancestors': _none,
   'base-uri': <String>["'self'"],